+++ src/Auth0.ManagementApi/Models/Actions/ActionBase.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Auth0.ManagementApi.Actions
+namespace Auth0.ManagementApi.Models.Actions
 {
     public abstract class ActionBase
     {
