+++ src/Auth0.ManagementApi/Models/Actions/ActionExecution.cs
@@ -1,15 +1,13 @@
 ﻿using System;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Auth0.ManagementApi.Actions
+namespace Auth0.ManagementApi.Models.Actions
 {
-    /// <summary>
-    /// Represents a Trigger Binding in Auth0
-    /// </summary>
-    public class TriggerBinding
+    public class ActionExecution
     {
         /// <summary>
-        /// The unique ID of this binding.
+        /// Identifies a specific execution.
         /// </summary>
         [JsonProperty("id")]
         public string Id { get; set; }
@@ -21,27 +19,27 @@ namespace Auth0.ManagementApi.Actions
         public string TriggerId { get; set; }
 
         /// <summary>
-        /// The connected action.
+        /// The overall status of an execution.
         /// </summary>
-        [JsonProperty("action")]
-        public Action Action { get; set; }
+        [JsonProperty("status")]
+        public string Status { get; set; }
 
         /// <summary>
-        /// The time when the binding was created.
+        /// Captures the results of a single action being executed.
+        /// </summary>
+        [JsonProperty("results")]
+        public IList<ActionExecutionResult> Results { get; set; }
+
+        /// <summary>
+        /// The time that the execution was started.
         /// </summary>
         [JsonProperty("created_at")]
         public DateTime CreatedAt { get; set; }
 
         /// <summary>
-        /// The time when the binding was updated.
+        /// The time that the execution finished executing.
         /// </summary>
         [JsonProperty("updated_at")]
         public DateTime UpdatedAt { get; set; }
-
-        /// <summary>
-        /// The name of the binding.
-        /// </summary>
-        [JsonProperty("display_name")]
-        public string DisplayName { get; set; }
     }
 }
