+++ src/Auth0.ManagementApi/Models/Actions/ActionSecret.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json;
 
-namespace Auth0.ManagementApi.Actions
+namespace Auth0.ManagementApi.Models.Actions
 {
     public class ActionSecret
     {
