+++ src/Auth0.ManagementApi/Models/Actions/ActionVersion.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Auth0.ManagementApi.Actions
+namespace Auth0.ManagementApi.Models.Actions
 {
     /// <summary>
     /// Represents a version for an action in Auth0
