+++ src/Auth0.ManagementApi/Models/Actions/UpdateTriggerBindingsRequest.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Auth0.ManagementApi.Actions
+namespace Auth0.ManagementApi.Models.Actions
 {
     /// <summary>
     /// Request configuration to update the actions that are bound (i.e. attached) to a trigger.
